+++ R/hlm_resid.R
@@ -198,7 +198,7 @@ hlm_resid.lmerMod <- function(object, level = 1, standardize = FALSE, include.ls
     
     # Assemble Tibble
     if (include.ls == TRUE) {
-      return.tbl <- tibble::tibble("row_id" = as.numeric(rownames(data)),
+      return.tbl <- tibble::tibble(".id" = as.numeric(rownames(data)),
                                    data,
                                    eb.resid,
                                    eb.fitted,
@@ -206,7 +206,7 @@ hlm_resid.lmerMod <- function(object, level = 1, standardize = FALSE, include.ls
                                    mar.fitted, 
                                    .name_repair = "minimal")
     } else { 
-      return.tbl <- tibble::tibble("row_id" = as.numeric(rownames(data)),
+      return.tbl <- tibble::tibble(".id" = as.numeric(rownames(data)),
                                    data,
                                    eb.resid,
                                    eb.fitted,
@@ -432,7 +432,7 @@ hlm_resid.lme <- function(object, level = 1, standardize = FALSE, include.ls = T
     
     # Continue to Assemble Tibble  
     if (include.ls == TRUE) {
-      return.tbl <- tibble::tibble("row_id" = as.numeric(rownames(model.data)),
+      return.tbl <- tibble::tibble(".id" = as.numeric(rownames(model.data)),
                                    model.data,
                                    eb.resid,
                                    eb.fitted,
@@ -441,7 +441,7 @@ hlm_resid.lme <- function(object, level = 1, standardize = FALSE, include.ls = T
                                    mar.fitted, 
                                    .name_repair = "minimal")
     } else { 
-      return.tbl <- tibble::tibble("row_id" = as.numeric(rownames(model.data)),
+      return.tbl <- tibble::tibble(".id" = as.numeric(rownames(model.data)),
                                    model.data,
                                    eb.resid,
                                    eb.fitted,