+++ NEWS.md
@@ -1,4 +1,4 @@
-# HLMdiag 0.5.9000
+# HLMdiag 0.5.1
 
 * Added a `NEWS.md` file to track changes to the package.
 * `cooks.distance` and `mdffits` with `include.attr = TRUE` now returns a tibble with column names for the fixed effects in agreement with the fixef() output.